Google Visualization fallback added
I added a fallback for the Google Visualization pie charts in the skills
section. The fallbacks are simply images of the charts taken from
screenshots. They are visible by default. When the charts are drawn
correctly, they fire a ready event. A handler was added for this event
to remove the images by setting their display to none.